solutions/7_不会ML/app/requirements.txt
rapidocr_paddle
rapidocr_onnxruntime==1.3.8
FlagEmbedding==1.2.9
pdfplumber==0.11.0
langchain==0.2.0
Pillow==9.2.0
pdf2image==1.17.0
pandas
numpy
modelscope==1.14.0
accelerate==0.27.2
deepspeed==0.13.4
einops
tiktoken==0.6.0
transformers==4.32.0
transformers-stream-generator==0.0.4
uvicorn==0.27.1
tokenizers==0.13.3
peft==0.4.0
faiss-cpu==1.8.0
deepspeed==0.13.4
transformers==4.39.1
fastapi==0.110.0
sentencepiece==0.1.99
jsonlines